Implement Detect
As the comment suggested, use github.com/ryanuber/go-license.

For this, to recognize MPL2, licenses/mpl2.header needed a small
change: instead of "v. 2.0", now it contains "version 2.0".
Hope this is ok.

For set, I've added a detectLicense + removeLicense before the
insertLicense, to mitigate license duplication.
This does work for licenses without blank lines, but for more
comples header formats, as only the first paragraph is removed.

So mpl2 -> gpl2 ok, but not gpl2 -> mpl2.